Learn NgFor
We use the NgFor directive to loop over an array of items and create multiple elements dynamically from a template element.

The template element is the element the directive is attached to.

We can nest muliple NgFor directives together.

We can get the index of the item we are looping over by assigning index to a variable in the NgFor expression.