Learn NgClass & NgStyle
Both the NgStyle and NgClass directives can be used to conditionally set the look and feel of your application.

NgStyle gives you fine grained control on individual properties. But if you want to make changes to multiple properties at once, creating a class which bundles those properties and adding the class with NgClass makes more sense.