Learn Model-Driven Forms
In this lecture we created a simple HTML form. We created a form model on our component using the FormGroup and FormControl classes.

Then by using directives such as formGroup, formControlName and formGroupName we linked our HTML form to our form model.

In the next lecture we will learn how to add validators to our forms to give visual feedback when the data entered is invalid.